+++ eqcorrscan/core/match_filter/tribe.py
@@ -20,14 +20,16 @@ import tarfile
 import tempfile
 import traceback
 import logging
+import multiprocessing
 
 import numpy as np
-from obspy import Catalog, Stream, read, read_events
+from obspy import Catalog, Stream, read
 from obspy.core.event import Comment, CreationInfo
+from concurrent.futures import ThreadPoolExecutor, ProcessPoolExecutor
 
+from eqcorrscan.core.match_filter.party import Party, _read_catalog_pass_error
 from eqcorrscan.core.match_filter.template import (
     Template, quick_group_templates)
-from eqcorrscan.core.match_filter.party import Party
 from eqcorrscan.core.match_filter.helpers import (
     _safemembers, _par_read, get_waveform_client)
 from eqcorrscan.core.match_filter.matched_filter import (
@@ -260,7 +262,8 @@ class Tribe(object):
         """
         return copy.deepcopy(self)
 
-    def write(self, filename, compress=True, catalog_format="QUAKEML"):
+    def write(self, filename, compress=True, catalog_format="QUAKEML",
+              max_events_per_file=1000):
         """
         Write the tribe to a file using tar archive formatting.
 
@@ -277,6 +280,11 @@ class Tribe(object):
             SC3ML, QUAKEML are supported. Note that not all information is
             written for all formats (QUAKEML is the most complete, but is
             slow for IO).
+        :type max_events_per_file: int
+        :param max_events_per_file:
+            Maximum events to write per QuakeML file. When this number is
+            exceeded a new file will be written to the tar archive - allows
+            for parallel IO.
 
         .. rubric:: Example
 
@@ -302,10 +310,16 @@ class Tribe(object):
                     elif comment.text.startswith("eqcorrscan_template_"):
                         comment.text = "eqcorrscan_template_{0}".format(t.name)
                 tribe_cat.append(t.event)
-        if len(tribe_cat) > 0:
-            tribe_cat.write(
-                os.path.join(dirname, 'tribe_cat.{0}'.format(
-                    CAT_EXT_MAP[catalog_format])), format=catalog_format)
+        i, j = 0, 0
+        while i < len(tribe_cat):
+            sub_catalog = Catalog(
+                tribe_cat[i:i + max_events_per_file])
+            sub_catalog.write(
+                os.path.join(dirname, 'tribe_cat.{0}.{1}'.format(
+                    j, CAT_EXT_MAP[catalog_format])),
+                format=catalog_format)
+            j += 1
+            i += max_events_per_file
         for template in self.templates:
             template.st.write(
                 os.path.join(dirname, '{0}.ms'.format(template.name)),
@@ -336,12 +350,16 @@ class Tribe(object):
                 parfile.write('\n')
         return self
 
-    def read(self, filename):
+    def read(self, filename, cores=1):
         """
         Read a tribe of templates from a tar formatted file.
 
         :type filename: str
         :param filename: File to read templates from.
+        :type cores: int
+        :param cores:
+            Maximum number of processes to read waveforms with. Defaults to
+            serial.
 
         .. rubric:: Example
 
@@ -356,25 +374,77 @@ class Tribe(object):
             temp_dir = tempfile.mkdtemp()
             arc.extractall(path=temp_dir, members=_safemembers(arc))
             tribe_dir = glob.glob(temp_dir + os.sep + '*')[0]
-            self._read_from_folder(dirname=tribe_dir)
+            self._read_from_folder(
+                dirname=tribe_dir, cores=cores)
         shutil.rmtree(temp_dir)
         return self
 
-    def _read_from_folder(self, dirname):
+    def _read_from_folder(self, dirname, cores=1):
         """
         Internal folder reader.
 
         :type dirname: str
         :param dirname: Folder to read from.
+        :type cores: int
+        :param cores:
+            Maximum number of processes to read waveforms with. Defaults to
+            serial.
         """
+        cores = cores or multiprocessing.cpu_count()
         templates = _par_read(dirname=dirname, compressed=False)
         t_files = glob.glob(dirname + os.sep + '*.ms')
-        tribe_cat_file = glob.glob(os.path.join(dirname, "tribe_cat.*"))
-        if len(tribe_cat_file) != 0:
-            tribe_cat = read_events(tribe_cat_file[0])
+        t_files_dict = {os.path.splitext(os.path.basename(t))[0]: t
+                        for t in t_files}
+        if len(t_files) > len(t_files_dict):
+            Logger.warning("Duplicate template files found")
+        tribe_cat_files = glob.glob(os.path.join(dirname, "tribe_cat.*"))
+        tribe_cat = Catalog()
+        Logger.info("Reading tribe catalog")
+        if cores == 1 or len(tribe_cat_files) == 1:
+            for tribe_cat_file in tribe_cat_files:
+                tribe_cat += _read_catalog_pass_error(tribe_cat_file)
         else:
-            tribe_cat = Catalog()
-        previous_template_names = [t.name for t in self.templates]
+            cores = min(cores, len(tribe_cat_files))
+            with ProcessPoolExecutor(max_workers=cores) as executor:
+                _cats = executor.map(
+                    _read_catalog_pass_error, tribe_cat_files,
+                    chunksize=(len(tribe_cat_files) // cores) + (
+                            len(tribe_cat_files) % cores))
+            for _cat in _cats:
+                tribe_cat += _cat
+
+        previous_template_names = {t.name for t in self.templates}
+        template_names = {t.name for t in templates}.difference(
+            previous_template_names)
+        assert template_names.issubset(set(t_files_dict.keys())), \
+            "Missing waveform files for templates"
+        template_names = list(template_names)  # Need an ordered object
+
+        Logger.info("Reading tribe streams")
+        if cores > 1 and len(template_names) > 1:
+            # Need to be able to link stream to template name
+            filenames = [t_files_dict.get(t_name)
+                         for t_name in template_names
+                         if t_name in t_files_dict.keys()]
+            # This should never fail due to above assertion...
+            assert len(filenames) == len(template_names), \
+                "Missing waveform files for templates"
+            # Don't use more processes than files
+            cores = min(cores, len(filenames))
+            t_files_dict_reverse_mapper = {
+                value: key for key, value in t_files_dict.items()}
+            template_streams = dict()
+            with ThreadPoolExecutor(max_workers=cores) as executor:
+                for t_file, st in zip(
+                        filenames, executor.map(read, filenames)):
+                    t_name = t_files_dict_reverse_mapper.get(t_file)
+                    template_streams.update({t_name: st})
+        else:
+            template_streams = {
+                key: read(value) for key, value in t_files_dict.items()
+                if key in template_names}
+
+        Logger.info("Reconstructing tribe")
         for template in templates:
             if template.name in previous_template_names:
                 # Don't read in for templates that we already have.
@@ -383,16 +453,11 @@ class Tribe(object):
                 for comment in event.comments:
                     if comment.text == 'eqcorrscan_template_' + template.name:
                         template.event = event
-            t_file = [t for t in t_files
-                      if t.split(os.sep)[-1] == template.name + '.ms']
-            if len(t_file) == 0:
+            template.st = template_streams.get(template.name, None)
+            if not template.st:
                 Logger.error('No waveform for template: ' + template.name)
-                templates.remove(template)
                 continue
-            elif len(t_file) > 1:
-                Logger.warning('Multiple waveforms found, using: ' + t_file[0])
-            template.st = read(t_file[0])
-        self.templates.extend(templates)
+        self.templates.extend([t for t in templates if t.st])
         return
 
     def cluster(self, method, **kwargs):
@@ -1073,15 +1138,20 @@ class Tribe(object):
         return self
 
 
-def read_tribe(fname):
+def read_tribe(fname, cores=1):
     """
     Read a Tribe of templates from a tar archive.
 
     :param fname: Filename to read from
+    :type cores: int
+    :param cores:
+        Maximum number of processes to read waveforms with. Defaults to
+        serial.
+
     :return: :class:`eqcorrscan.core.match_filter.Tribe`
     """
     tribe = Tribe()
-    tribe.read(filename=fname)
+    tribe.read(filename=fname, cores=cores)
     return tribe