+++ CHANGES.md
@@ -1,4 +1,10 @@
 ## Current
+* core.match_filter.party:
+ - Implement parallel reading, and chunked writing to save memory on writing and 
+   accelerate reading - PR: #462
+* core.match_filter.tribe:
+ - Implement parallel reading and chunked writing to save memory on writing and
+   accelerate reading - PR: #462
 * core.match_filter.matched_filter
   - 5x speed up for MAD threshold calculation with parallel (threaded) MAD 
     calculation (#531).